+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 ##Team member: Castiel Li (CL) , Jason Chen (JC) , Harrison Milbradt (HM) , Mu Chen (Paul) Li (ML), Nate Chiang (NC)
 
 ## [0.2.3]
+### 2018-04-05
+#### Changed
+* Clean up migrations
+* Fix boat model
+* Fix reservation model
+* Fix reservation mvc
+
+## [0.2.3]
 ### 2018-04-02
 #### Changed
 * Fix boat seeding data - JC